requirements.txt
# Core dependencies
pydantic>=2.0.0
pyyaml>=6.0
colorama>=0.4.6
click>=8.1.0
chardet>=5.0.0

# MCP integration
mcp>=1.0.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0
hypothesis>=6.82.0

# Development dependencies
black>=23.7.0
isort>=5.12.0
mypy>=1.5.0
flake8>=6.0.0